# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  accounts.PasswordResetConfirmInput:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  accounts.PasswordResetRequestInput:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
  accounts.ResendVerificationInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing trip plans, hops, and stays
  title: Trip Planner API
  version: "1.0"
paths:
...
  /auth/password-reset/confirm:
    post:
      consumes:
      - application/json
      description: Validates the reset token and updates the password.
      parameters:
      - description: Token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.PasswordResetConfirmInput'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid or expired token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset a user's password
      tags:
      - auth
  /auth/password-reset/request:
    post:
      consumes:
      - application/json
      description: Sends a password reset link to the given email. Always returns
        200 to prevent email enumeration.
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.PasswordResetRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request a password reset email
      tags:
      - auth
...
  /auth/verify-email:
    get:
      description: Marks the user's email as verified using the token sent at signup.
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify email address
      tags:
      - auth
  /auth/verify-email/resend:
    post:
      consumes:
      - application/json
      description: Regenerates and resends the email verification link.
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.ResendVerificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Resend email verification
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"